+++ src/approx/inverse_hyperbolic.test.ts
@@ -1,59 +1,37 @@
-import { RoundingMode } from '../types';
-import { ExactNumber } from '../ExactNumber';
 import { acosh, asinh, atanh } from './inverse_hyperbolic';
+import { compareError, testStability } from '../testHelper.test';
 
 describe('hyperbolic', () => {
   it('asinh', () => {
-    const range = [-4, 4];
-    for (let i = range[0]; i <= range[1]; i += 0.004) {
+    for (let i = -4; i <= 4; i += 0.004) {
       const jsResult = Math.asinh(i).toString();
-      const jsRounded = ExactNumber(jsResult).round(10, RoundingMode.TO_ZERO).toFixed(10);
-      const exactResult = ExactNumber(asinh(i.toString(), 10)).toFixed(10);
-      expect(exactResult).toBe(jsRounded);
+      compareError(asinh(i.toString(), 30), jsResult);
     }
   });
 
   it('asinh 7/12', () => {
-    const ASINH_7_OVER_12 = asinh('7/12', 150);
-
-    for (let i = 1; i < 150; i++) {
-      expect(asinh('7/12', i)).toBe(ASINH_7_OVER_12.slice(0, i + 2));
-    }
+    testStability(decimals => asinh('7/12', decimals), 150);
   });
 
   it('acosh', () => {
-    const range = [1, 8];
-    for (let i = range[0]; i <= range[1]; i += 0.004) {
+    for (let i = 1; i <= 8; i += 0.004) {
       const jsResult = Math.acosh(i).toString();
-      const jsRounded = ExactNumber(jsResult).round(10, RoundingMode.TO_ZERO).toFixed(10);
-      const exactResult = ExactNumber(acosh(i.toString(), 10)).toFixed(10);
-      expect(exactResult).toBe(jsRounded);
+      compareError(acosh(i.toString(), 30), jsResult);
     }
   });
 
   it('acosh 17/12', () => {
-    const ACOSH_17_OVER_12 = acosh('17/12', 150);
-
-    for (let i = 1; i < 150; i++) {
-      expect(acosh('17/12', i)).toBe(ACOSH_17_OVER_12.slice(0, i + 2));
-    }
+    testStability(decimals => acosh('17/12', decimals), 150);
   });
 
   it('atanh', () => {
-    const range = [-0.9999, 0.9999];
-    for (let i = range[0]; i <= range[1]; i += 0.002) {
+    for (let i = -0.9999; i <= 0.9999; i += 0.002) {
       const jsResult = Math.atanh(i).toString();
-      const jsRounded = ExactNumber(jsResult).round(10, RoundingMode.TO_ZERO).toFixed(10);
-      const exactResult = ExactNumber(atanh(i.toString(), 10)).toFixed(10);
-      expect(exactResult).toBe(jsRounded);
+      compareError(atanh(i.toString(), 30), jsResult);
     }
   });
 
   it('atanh 7/12', () => {
-    const ATANH_7_OVER_12 = atanh('7/12', 150);
-
-    for (let i = 1; i < 150; i++) {
-      expect(atanh('7/12', i)).toBe(ATANH_7_OVER_12.slice(0, i + 2));
-    }
+    testStability(decimals => atanh('7/12', decimals), 150);
   });
 });