+++ src/approx/inverse_trigonometry.ts
@@ -3,36 +3,15 @@ import { ExactNumberType } from '../types';
 import { sqrt } from './roots';
 import { PI } from './trigonometry';
 
-export const atan = (value: number | bigint | string | ExactNumberType, digits: number) => {
-  let x = ExactNumber(value);
-
-  if (x.isZero()) return '0';
-  if (x.abs().isOne()) {
-    return ExactNumber(PI(digits))
-      .div(4 * x.sign())
-      .toFixed(digits);
-  }
-
-  // Ensure |x| < 0.42
-  // atan(x) = 2 * atan(x / (1 + sqrt(1 + x^2)))
-  let reductionSteps = 0;
-  while (x.abs().gt('0.42')) {
-    const root = ExactNumber(sqrt(x.pow(2n).add(1n), digits + 10));
-    x = x.div(root.add(1n));
-    reductionSteps++;
-  }
-
-  // atan(x) = x - x^3/3 + x^5/5 - x^7/7 + ...
-  const x2 = x.pow(2).normalize();
-  const x4 = x2.pow(2).normalize();
-
+// atan x = x - x^3/3 + x^5/5 - x^7/7 + ...
+function* atanGenerator(x: ExactNumberType, decimals: number) {
+  const x2 = x.pow(2n).normalize();
+  const x4 = x2.pow(2n).normalize();
   let denominator = 3n;
 
-  let xk = x.sub(x.mul(x2).div(denominator)) as ExactNumberType;
+  let sum = x.sub(x.mul(x2).div(denominator));
 
   let xPow = x.mul(x4);
-
-  const maxError = ExactNumber(1n).div(10n ** BigInt(digits + 10));
 
   while (true) {
     // x^5/d - x^7/(d + 2)
@@ -44,28 +23,55 @@ export const atan = (value: number | bigint | string | ExactNumberType, digits: 
     denominator = denominator2;
     xPow = xPow.mul(x4);
 
-    xk = xk.add(term).trunc(digits + 10);
+    sum = sum.add(term).trunc(decimals + 10);
 
+    yield { term, sum };
+  }
+}
+
+export const atan = (value: number | bigint | string | ExactNumberType, decimals: number): ExactNumberType => {
+  let x = ExactNumber(value);
+
+  if (x.isZero()) return ExactNumber(0);
+  if (x.abs().isOne()) {
+    return ExactNumber(PI(decimals))
+      .div(4 * x.sign())
+      .trunc(decimals);
+  }
+
+  // Ensure |x| < 0.42
+  // atan(x) = 2 * atan(x / (1 + sqrt(1 + x^2)))
+  let reductionSteps = 0;
+  const reductionLimit = ExactNumber('0.42');
+  while (x.abs().gt(reductionLimit)) {
+    const root = ExactNumber(sqrt(x.pow(2n).add(1n), decimals + 10));
+    x = x.div(root.add(1n));
+    reductionSteps++;
+  }
+
+  const maxError = ExactNumber(`1e-${decimals + 10}`);
+
+  const gen = atanGenerator(x, decimals);
+  for (const { term, sum } of gen) {
     if (term.abs().lt(maxError)) {
-      break;
+      // undo argument reduction
+      const res = sum.mul(2n ** BigInt(reductionSteps));
+      return res.trunc(decimals);
     }
   }
 
-  // undo argument reduction
-  xk = xk.mul(2n ** BigInt(reductionSteps));
-
-  return xk.toFixed(digits);
+  return ExactNumber(0);
 };
 
-export const asin = (value: number | bigint | string | ExactNumberType, digits: number): string => {
+export const asin = (value: number | bigint | string | ExactNumberType, decimals: number): ExactNumberType => {
   const x = ExactNumber(value);
 
-  if (x.isZero()) return '0';
+  if (x.isZero()) return ExactNumber(0);
   if (x.abs().isOne()) {
-    return ExactNumber(PI(digits)).mul(x.sign()).div(2n).toFixed(digits);
+    return ExactNumber(PI(decimals)).mul(x.sign()).div(2n).trunc(decimals);
   }
   if (x.abs().eq('1/2')) {
-    return ExactNumber(PI(digits)).mul(x.sign()).div(6n).toFixed(digits);
+    return ExactNumber(PI(decimals)).mul(x.sign()).div(6n).trunc(decimals);
   }
   if (x.gt(1n) || x.lt(-1n)) {
     throw new Error('Out of range');
@@ -73,27 +79,27 @@ export const asin = (value: number | bigint | string | ExactNumberType, digits: 
 
   // asin(x) = 2*atan(x / (1 + sqrt(1 - x^2)))
 
-  const root = ExactNumber(sqrt(x.pow(2n).neg().add(1), digits + 10));
-  const atangent = ExactNumber(atan(x.div(root.add(1n)), digits + 10));
-  return atangent.mul(2).toFixed(digits);
+  const root = ExactNumber(sqrt(x.pow(2n).neg().add(1), decimals + 10));
+  const atangent = ExactNumber(atan(x.div(root.add(1n)), decimals + 10));
+  return atangent.mul(2).trunc(decimals);
 };
 
-export const acos = (value: number | bigint | string | ExactNumberType, digits: number): string => {
+export const acos = (value: number | bigint | string | ExactNumberType, decimals: number): ExactNumberType => {
   const x = ExactNumber(value);
 
-  if (x.isZero()) return ExactNumber(PI(digits)).div(2n).toFixed(digits);
+  if (x.isZero()) return ExactNumber(PI(decimals)).div(2n).trunc(decimals);
 
   if (x.isOne()) {
-    return '0';
+    return ExactNumber(0);
   }
 
   if (x.abs().isOne()) {
-    return PI(digits);
+    return PI(decimals);
   }
 
   if (x.abs().eq('1/2')) {
-    const PI_OVER_3 = ExactNumber(PI(digits)).div(3n);
-    return x.sign() === -1 ? PI_OVER_3.mul(2n).toFixed(digits) : PI_OVER_3.toFixed(digits);
+    const PI_OVER_3 = ExactNumber(PI(decimals)).div(3n);
+    return x.sign() === -1 ? PI_OVER_3.mul(2n).trunc(decimals) : PI_OVER_3.trunc(decimals);
   }
 
   if (x.gt(1n) || x.lt(-1n)) {
@@ -101,8 +107,8 @@ export const acos = (value: number | bigint | string | ExactNumberType, digits: 
   }
 
   // acos(x) = pi/2 - asin(x)
-  return ExactNumber(PI(digits + 10))
+  return ExactNumber(PI(decimals + 10))
     .div(2n)
-    .sub(asin(x, digits + 10))
-    .toFixed(digits);
+    .sub(asin(x, decimals + 10))
+    .trunc(decimals);
 };